+++ motep/potentials/mtp/numba/engine.py
@@ -9,7 +9,7 @@ from motep.potentials.mtp import get_types
 from motep.potentials.mtp.base import EngineBase
 
 from .chebyshev import _nb_calc_radial_basis, _nb_calc_radial_funcs
-from .utils import (
+from .moment import (
     _nb_calc_local_energy_and_gradient,
     _nb_calc_moment,
     _nb_forces_from_gradient,
